chore(parser): refactor envelope imports
