--- packages/parser/src/envelopes/index.ts
@@ -1,14 +1,0 @@
-export { ApiGatewayEnvelope } from './apigw.js';
-export { ApiGatewayV2Envelope } from './apigwv2.js';
-export { CloudWatchEnvelope } from './cloudwatch.js';
-export { DynamoDBStreamEnvelope } from './dynamodb.js';
-export { EventBridgeEnvelope } from './event-bridge.js';
-export { KafkaEnvelope } from './kafka.js';
-export { KinesisEnvelope } from './kinesis.js';
-export { KinesisFirehoseEnvelope } from './kinesis-firehose.js';
-export { LambdaFunctionUrlEnvelope } from './lambda.js';
-export { SnsEnvelope } from './sns.js';
-export { SqsEnvelope } from './sqs.js';
-export { SnsSqsEnvelope } from './snssqs.js';
-export { VpcLatticeEnvelope } from './vpc-lattice.js';
-export { VpcLatticeV2Envelope } from './vpc-latticev2.js';
