+++ packages/parser/tests/unit/envelopes/cloudwatch.test.ts
@@ -2,7 +2,7 @@ import { gunzipSync, gzipSync } from 'node:zlib';
 import { describe, expect, it } from 'vitest';
 import { ZodError, z } from 'zod';
 import { ParseError } from '../../../src';
-import { CloudWatchEnvelope } from '../../../src/envelopes/index.js';
+import { CloudWatchEnvelope } from '../../../src/envelopes/cloudwatch.js';
 import { DecompressError } from '../../../src/errors.js';
 import { JSONStringified } from '../../../src/helpers.js';
 import { getTestEvent } from '../helpers/utils.js';
