+++ packages/parser/tests/unit/envelopes/event-bridge.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { ZodError, z } from 'zod';
-import { EventBridgeEnvelope } from '../../../src/envelopes/index.js';
+import { EventBridgeEnvelope } from '../../../src/envelopes/eventbridge.js';
 import { ParseError } from '../../../src/errors.js';
 import type { EventBridgeEvent } from '../../../src/types/schema.js';
 import { getTestEvent, omit } from '../helpers/utils.js';
