+++ packages/parser/tests/unit/envelopes/kafka.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { ZodError, z } from 'zod';
-import { KafkaEnvelope } from '../../../src/envelopes/index.js';
+import { KafkaEnvelope } from '../../../src/envelopes/kafka.js';
 import { ParseError } from '../../../src/errors.js';
 import { JSONStringified } from '../../../src/helpers.js';
 import { getTestEvent } from '../helpers/utils.js';
