+++ packages/parser/tests/unit/envelopes/lambda.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { ZodError, z } from 'zod';
 import { ParseError } from '../../../src';
-import { LambdaFunctionUrlEnvelope } from '../../../src/envelopes/index.js';
+import { LambdaFunctionUrlEnvelope } from '../../../src/envelopes/lambda.js';
 import { JSONStringified } from '../../../src/helpers';
 import type { LambdaFunctionUrlEvent } from '../../../src/types';
 import { getTestEvent, omit } from '../helpers/utils.js';
