+++ packages/parser/tests/unit/envelopes/vpc-latticev2.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { ZodError, z } from 'zod';
-import { VpcLatticeV2Envelope } from '../../../src/envelopes/index.js';
+import { VpcLatticeV2Envelope } from '../../../src/envelopes/vpc-latticev2.js';
 import { ParseError } from '../../../src/errors.js';
 import { JSONStringified } from '../../../src/helpers.js';
 import type { VpcLatticeEventV2 } from '../../../src/types/index.js';
